backend/pyproject.toml
[tool.poetry]
name = "ph-rd-tt-web"
version = "0.1.0"
description = "This will be used to manage the dependency for project."
authors = ["Shrijeet-Srivastava <106809277+Shrijeet-Srivastava@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
requests = "^2.32.3"
boto3 = "^1.35.64"
mangum = "^0.19.0"
pandas = "^2.2.3"
jwt = "^1.3.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
pyarrow = "^18.0.0"
aws-lambda-powertools = "^3.3.0"
pyodbc = "5.2.0"
msal = "1.31.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
